+++ CHANGELOG.md
@@ -1,3 +1,33 @@
+<a name="9.7.0"></a>
+# [9.7.0](https://github.com/alfa-laboratory/arui-scripts/compare/v9.6.0...v9.7.0) (2020-10-22)
+
+
+### Features
+
+* **webpack:** split dynamic import from node_modules to chunk ([ba6a286](https://github.com/alfa-laboratory/arui-scripts/commit/ba6a286))
+
+
+
+<a name="9.6.0"></a>
+# [9.6.0](https://github.com/alfa-laboratory/arui-scripts/compare/v9.5.0...v9.6.0) (2020-10-08)
+
+
+### Features
+
+* **postcss-config:** add keepCssVars, revert plugin ([2d8a0d6](https://github.com/alfa-laboratory/arui-scripts/commit/2d8a0d6))
+
+
+
+<a name="9.5.0"></a>
+# [9.5.0](https://github.com/alfa-laboratory/arui-scripts/compare/v9.4.1...v9.5.0) (2020-09-22)
+
+
+### Features
+
+* **webpack:** replace style tags with optimized css bundle in dev ([22681f2](https://github.com/alfa-laboratory/arui-scripts/commit/22681f2))
+
+
+
 <a name="9.4.1"></a>
 ## [9.4.1](https://github.com/alfa-laboratory/arui-scripts/compare/v9.4.0...v9.4.1) (2020-09-11)